@@ -1,7 +1,3 @@
-func _idToString(_ s: String) -> String {
-    return StringDumper(s).dump()
-}
-
 public struct KDLNode: Equatable, CustomStringConvertible {
     var name: String
     var arguments: [KDLValue] = []
@@ -9,14 +5,16 @@ public struct KDLNode: Equatable, CustomStringConvertible {
     var keys: [String] = []
     var children: [KDLNode] = []
     var type: String? = nil
+    var version: UInt
 
-    public init(_ name: String, arguments: [KDLValue] = [], properties: [String:KDLValue] = [:], children: [KDLNode] = [], type: String? = nil) {
+    public init(_ name: String, arguments: [KDLValue] = [], properties: [String:KDLValue] = [:], children: [KDLNode] = [], type: String? = nil, version: UInt = 2) {
         self.name = name
         self.arguments = arguments
         self.properties = properties
         self.keys = Array(properties.keys)
         self.children = children
         self.type = type
+        self.version = version
     }
 
     public subscript(index: Int) -> KDLValue {
@@ -109,5 +107,12 @@ public struct KDLNode: Equatable, CustomStringConvertible {
             lhs.properties == rhs.properties &&
             lhs.children == rhs.children &&
             lhs.type == rhs.type
+    }
+
+    func _idToString(_ s: String) -> String {
+        switch version {
+        case 1: return StringDumperV1(s).dump()
+        default: return StringDumper(s).dump()
+        }
     }
 }
