@@ -2,10 +2,10 @@ import Testing
 import BigDecimal
 @testable import KDL
 
-@Suite("Tokenizer tests")
-final class KDLTokenizerTests {
+@Suite("Tokenizer v1 tests")
+final class KDLTokenizerV1Tests {
     @Test func testPeekAndPeekAfterNext() throws {
-        let tokenizer = KDLTokenizer("node 1 2 3")
+        let tokenizer = KDLTokenizerV1("node 1 2 3")
         try #expect(tokenizer.peekToken() == .IDENT("node"))
         try #expect(tokenizer.peekTokenAfterNext() == .WS)
         try #expect(tokenizer.nextToken() == .IDENT("node"))
@@ -14,100 +14,78 @@ final class KDLTokenizerTests {
     }
 
     @Test func testIdentifier() throws {
-        try #expect(KDLTokenizer("foo").nextToken() == .IDENT("foo"))
-        try #expect(KDLTokenizer("foo-bar123").nextToken() == .IDENT("foo-bar123"))
-        try #expect(KDLTokenizer("-").nextToken() == .IDENT("-"))
-        try #expect(KDLTokenizer("--").nextToken() == .IDENT("--"))
+        try #expect(KDLTokenizerV1("foo").nextToken() == .IDENT("foo"))
+        try #expect(KDLTokenizerV1("foo-bar123").nextToken() == .IDENT("foo-bar123"))
     }
 
     @Test func testString() throws {
-        try #expect(KDLTokenizer(#""foo""#).nextToken() == .STRING("foo"))
-        try #expect(KDLTokenizer(#""foo\nbar""#).nextToken() == .STRING("foo\nbar"))
-        try #expect(KDLTokenizer(#""\u{10FFF}""#).nextToken() == .STRING("\u{10FFF}"))
-        try #expect(KDLTokenizer("\"\\\n\n\nfoo\"").nextToken() == .STRING("foo"))
+        try #expect(KDLTokenizerV1(#""foo""#).nextToken() == .STRING("foo"))
+        try #expect(KDLTokenizerV1(#""foo\nbar""#).nextToken() == .STRING("foo\nbar"))
+        try #expect(KDLTokenizerV1(#""\u{10FFF}""#).nextToken() == .STRING("\u{10FFF}"))
     }
 
     @Test func testRawstring() throws {
-        try #expect(KDLTokenizer(##"#"foo\nbar"#"##).nextToken() == .RAWSTRING(#"foo\nbar"#))
-        try #expect(KDLTokenizer(##"#"foo"bar"#"##).nextToken() == .RAWSTRING(#"foo"bar"#))
-        try #expect(KDLTokenizer(###"##"foo"#bar"##"###).nextToken() == .RAWSTRING(##"foo"#bar"##))
-        try #expect(KDLTokenizer(##"#""foo""#"##).nextToken() == .RAWSTRING(#""foo""#))
+        try #expect(KDLTokenizerV1(##"r"foo\nbar""##).nextToken() == .RAWSTRING(#"foo\nbar"#))
+        try #expect(KDLTokenizerV1(##"r#"foo"bar"#"##).nextToken() == .RAWSTRING(#"foo"bar"#))
+        try #expect(KDLTokenizerV1(###"r##"foo"#bar"##"###).nextToken() == .RAWSTRING(##"foo"#bar"##))
+        try #expect(KDLTokenizerV1(##"r#""foo""#"##).nextToken() == .RAWSTRING(#""foo""#))
 
-        var tokenizer = KDLTokenizer(##"node #"C:\Users\zkat\"#"##)
+        var tokenizer = KDLTokenizerV1(##"node r"C:\Users\zkat\""##)
         try #expect(tokenizer.nextToken() == .IDENT("node"))
         try #expect(tokenizer.nextToken() == .WS)
         try #expect(tokenizer.nextToken() == .RAWSTRING(#"C:\Users\zkat\"#))
 
-        tokenizer = KDLTokenizer(##"other-node #"hello"world"#"##)
+        tokenizer = KDLTokenizerV1(##"other-node r#"hello"world"#"##)
         try #expect(tokenizer.nextToken() == .IDENT("other-node"))
         try #expect(tokenizer.nextToken() == .WS)
         try #expect(tokenizer.nextToken() == .RAWSTRING(#"hello"world"#))
     }
 
     @Test func testInteger() throws {
-        try #expect(KDLTokenizer("123").nextToken() == .INTEGER(123))
-        try #expect(KDLTokenizer("0x0123456789abcdef").nextToken() == .INTEGER(0x0123456789abcdef))
-        try #expect(KDLTokenizer("0o01234567").nextToken() == .INTEGER(0o01234567))
-        try #expect(KDLTokenizer("0b101001").nextToken() == .INTEGER(0b101001))
-        try #expect(KDLTokenizer("-0x0123456789abcdef").nextToken() == .INTEGER(-0x0123456789abcdef))
-        try #expect(KDLTokenizer("-0o01234567").nextToken() == .INTEGER(-0o01234567))
-        try #expect(KDLTokenizer("-0b101001").nextToken() == .INTEGER(-0b101001))
-        try #expect(KDLTokenizer("+0x0123456789abcdef").nextToken() == .INTEGER(0x0123456789abcdef))
-        try #expect(KDLTokenizer("+0o01234567").nextToken() == .INTEGER(0o01234567))
-        try #expect(KDLTokenizer("+0b101001").nextToken() == .INTEGER(0b101001))
+        try #expect(KDLTokenizerV1("123").nextToken() == .INTEGER(123))
+        try #expect(KDLTokenizerV1("0x0123456789abcdef").nextToken() == .INTEGER(0x0123456789abcdef))
+        try #expect(KDLTokenizerV1("0o01234567").nextToken() == .INTEGER(0o01234567))
+        try #expect(KDLTokenizerV1("0b101001").nextToken() == .INTEGER(0b101001))
+        try #expect(KDLTokenizerV1("-0x0123456789abcdef").nextToken() == .INTEGER(-0x0123456789abcdef))
+        try #expect(KDLTokenizerV1("-0o01234567").nextToken() == .INTEGER(-0o01234567))
+        try #expect(KDLTokenizerV1("-0b101001").nextToken() == .INTEGER(-0b101001))
+        try #expect(KDLTokenizerV1("+0x0123456789abcdef").nextToken() == .INTEGER(0x0123456789abcdef))
+        try #expect(KDLTokenizerV1("+0o01234567").nextToken() == .INTEGER(0o01234567))
+        try #expect(KDLTokenizerV1("+0b101001").nextToken() == .INTEGER(0b101001))
     }
 
     @Test func testFloat() throws {
-        try #expect(KDLTokenizer("1.23").nextToken() == .DECIMAL(BigDecimal("1.23")))
-        try #expect(KDLTokenizer("#inf").nextToken() == .FLOAT(Float.infinity))
-        try #expect(KDLTokenizer("#-inf").nextToken() == .FLOAT(-Float.infinity))
-        let nan = try KDLTokenizer("#nan").nextToken()
-        switch nan {
-            case .FLOAT(let x): #expect(x.isNaN)
-            default: Issue.record("token was not a .FLOAT")
-        }
+        try #expect(KDLTokenizerV1("1.23").nextToken() == .DECIMAL(BigDecimal("1.23")))
     }
 
     @Test func testBooleazn() throws {
-        try #expect(KDLTokenizer("#true").nextToken() == .TRUE)
-        try #expect(KDLTokenizer("#false").nextToken() == .FALSE)
+        try #expect(KDLTokenizerV1("true").nextToken() == .TRUE)
+        try #expect(KDLTokenizerV1("false").nextToken() == .FALSE)
     }
 
     @Test func testNull() throws {
-        try #expect(KDLTokenizer("#null").nextToken() == .NULL)
+        try #expect(KDLTokenizerV1("null").nextToken() == .NULL)
     }
 
     @Test func testSymbols() throws {
-        try #expect(KDLTokenizer("{").nextToken() == .LBRACE)
-        try #expect(KDLTokenizer("}").nextToken() == .RBRACE)
-    }
-
-    @Test func testEquals() throws {
-        try #expect(KDLTokenizer("=").nextToken() == .EQUALS)
-        try #expect(KDLTokenizer(" =").nextToken() == .EQUALS)
-        try #expect(KDLTokenizer("= ").nextToken() == .EQUALS)
-        try #expect(KDLTokenizer(" = ").nextToken() == .EQUALS)
-        try #expect(KDLTokenizer(" =foo").nextToken() == .EQUALS)
+        try #expect(KDLTokenizerV1("{").nextToken() == .LBRACE)
+        try #expect(KDLTokenizerV1("}").nextToken() == .RBRACE)
+        try #expect(KDLTokenizerV1("=").nextToken() == .EQUALS)
     }
 
     @Test func testWhitespace() throws {
-        try #expect(KDLTokenizer(" ").nextToken() == .WS)
-        try #expect(KDLTokenizer("\t").nextToken() == .WS)
-        try #expect(KDLTokenizer("    \t").nextToken() == .WS)
-        try #expect(KDLTokenizer("\\\n").nextToken() == .WS)
-        try #expect(KDLTokenizer("\\").nextToken() == .WS)
-        try #expect(KDLTokenizer("\\//some comment\n").nextToken() == .WS)
-        try #expect(KDLTokenizer("\\ //some comment\n").nextToken() == .WS)
-        try #expect(KDLTokenizer("\\//some comment").nextToken() == .WS)
-        try #expect(KDLTokenizer(" \\\n").nextToken() == .WS)
-        try #expect(KDLTokenizer(" \\//some comment\n").nextToken() == .WS)
-        try #expect(KDLTokenizer(" \\ //some comment\n").nextToken() == .WS)
-        try #expect(KDLTokenizer(" \\//some comment").nextToken() == .WS)
-        try #expect(KDLTokenizer(" \\\n  \\\n  ").nextToken() == .WS)
+        try #expect(KDLTokenizerV1(" ").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("\t").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("    \t").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("\\\n").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("\\").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("\\//some comment\n").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("\\ //some comment\n").nextToken() == .WS)
+        try #expect(KDLTokenizerV1("\\//some comment").nextToken() == .WS)
     }
 
     @Test func testMultipleTokens() throws {
-        let tokenizer = KDLTokenizer("node 1 \"two\" a=3")
+        let tokenizer = KDLTokenizerV1("node 1 \"two\" a=3")
 
         try #expect(tokenizer.nextToken() == .IDENT("node"))
         try #expect(tokenizer.nextToken() == .WS)
@@ -123,9 +101,9 @@ final class KDLTokenizerTests {
     }
 
     @Test func testSingleLineComment() throws {
-        try #expect(KDLTokenizer("// comment").nextToken() == .EOF)
+        try #expect(KDLTokenizerV1("// comment").nextToken() == .EOF)
 
-        let tokenizer = KDLTokenizer("""
+        let tokenizer = KDLTokenizerV1("""
         node1
         // comment
         node2
@@ -140,7 +118,7 @@ final class KDLTokenizerTests {
     }
 
     @Test func testMultilineComment() throws {
-        let tokenizer = KDLTokenizer("foo /*bar=1*/ baz=2");
+        let tokenizer = KDLTokenizerV1("foo /*bar=1*/ baz=2");
 
         try #expect(tokenizer.nextToken() == .IDENT("foo"))
         try #expect(tokenizer.nextToken() == .WS)
@@ -152,13 +130,13 @@ final class KDLTokenizerTests {
     }
 
     @Test func testUtf8() throws {
-        try #expect(KDLTokenizer("😁").nextToken() == .IDENT("😁"))
-        try #expect(KDLTokenizer(#""😁""#).nextToken() == .STRING("😁"))
-        try #expect(KDLTokenizer("ノード").nextToken() == .IDENT("ノード"))
-        try #expect(KDLTokenizer("お名前").nextToken() == .IDENT("お名前"))
-        try #expect(KDLTokenizer(#""☜(ﾟヮﾟ☜)""#).nextToken() == .STRING("☜(ﾟヮﾟ☜)"))
+        try #expect(KDLTokenizerV1("😁").nextToken() == .IDENT("😁"))
+        try #expect(KDLTokenizerV1(#""😁""#).nextToken() == .STRING("😁"))
+        try #expect(KDLTokenizerV1("ノード").nextToken() == .IDENT("ノード"))
+        try #expect(KDLTokenizerV1("お名前").nextToken() == .IDENT("お名前"))
+        try #expect(KDLTokenizerV1(#""☜(ﾟヮﾟ☜)""#).nextToken() == .STRING("☜(ﾟヮﾟ☜)"))
 
-        let tokenizer = KDLTokenizer("""
+        let tokenizer = KDLTokenizerV1("""
         smile "😁"
         ノード お名前="☜(ﾟヮﾟ☜)"
         """);
@@ -177,7 +155,7 @@ final class KDLTokenizerTests {
     }
 
     @Test func testSemicolon() throws {
-        let tokenizer = KDLTokenizer("node1; node2");
+        let tokenizer = KDLTokenizerV1("node1; node2");
 
         try #expect(tokenizer.nextToken() == .IDENT("node1"))
         try #expect(tokenizer.nextToken() == .SEMICOLON)
@@ -188,7 +166,7 @@ final class KDLTokenizerTests {
     }
 
     @Test func testSlashdash() throws {
-        let tokenizer = KDLTokenizer("""
+        let tokenizer = KDLTokenizerV1("""
         /-mynode /-"foo" /-key=1 /-{
             a
         }
@@ -217,7 +195,7 @@ final class KDLTokenizerTests {
     }
 
     @Test func testMultilineNodes() throws {
-        let tokenizer = KDLTokenizer("""
+        let tokenizer = KDLTokenizerV1("""
         title \\
             "Some title"
         """)
@@ -230,22 +208,21 @@ final class KDLTokenizerTests {
     }
 
     @Test func testTypes() throws {
-        var tokenizer = KDLTokenizer("(foo)bar")
+        var tokenizer = KDLTokenizerV1("(foo)bar")
         try #expect(tokenizer.nextToken() == .LPAREN)
         try #expect(tokenizer.nextToken() == .IDENT("foo"))
         try #expect(tokenizer.nextToken() == .RPAREN)
         try #expect(tokenizer.nextToken() == .IDENT("bar"))
 
-        tokenizer = KDLTokenizer("(foo)/*asdf*/bar")
+        tokenizer = KDLTokenizerV1("(foo)/*asdf*/bar")
         try #expect(tokenizer.nextToken() == .LPAREN)
         try #expect(tokenizer.nextToken() == .IDENT("foo"))
         try #expect(tokenizer.nextToken() == .RPAREN)
-        try #expect(tokenizer.nextToken() == .IDENT("bar"))
+        #expect(throws: (any Error).self) { try tokenizer.nextToken() }
 
-        tokenizer = KDLTokenizer("(foo/*asdf*/)bar")
+        tokenizer = KDLTokenizerV1("(foo/*asdf*/)bar")
         try #expect(tokenizer.nextToken() == .LPAREN)
         try #expect(tokenizer.nextToken() == .IDENT("foo"))
-        try #expect(tokenizer.nextToken() == .RPAREN)
-        try #expect(tokenizer.nextToken() == .IDENT("bar"))
+        #expect(throws: (any Error).self) { try tokenizer.nextToken() }
     }
 }
